requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
transformers==4.30.0
torch==2.4.0
regex==2023.6.3  
python-dotenv==1.0.0  
